feat: теты контроллеров разбиты на да класса
